% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_observed.R
\name{info_observed}
\alias{info_observed}
\title{Calculate the information matrix (based on the observed likelihood)}
\usage{
info_observed(model_fit, nqmc = 2000)
}
\arguments{
\item{model_fit}{a model fit returned by \code{joint_multistate_model},}

\item{nqmc}{an integer value. \cr 
Number of quasi Monte Carlo points for evaluating numeric integration.}
}
\value{
a list containing the information matrices
}
\description{
Calculate the information matrix (based on the observed likelihood)
}
