% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_multistate_model.R
\name{joint_multistate_model}
\alias{joint_multistate_model}
\title{Fit a joint model of multivariate longitudinal data and multistate data}
\usage{
joint_multistate_model(
  yyikj,
  ttikj,
  nobsik,
  zzi,
  xxi,
  statebegin,
  stateend,
  timebegin,
  timeend,
  ntransition,
  possible_transitions,
  basismatrix_long,
  nqmc = 2000
)
}
\arguments{
\item{yyikj}{3-dimensional numeric array of longitudinal data. \cr
\code{yyikj[i,k,j]} is the jth observation of the kth longitudinal measure of the ith subject.}

\item{ttikj}{3-dimensional integer array of longitudinal observation times. \cr
\code{ttikj[i,k,j]} is the jth longitudinal observation time of the kth longitudinal measure of the ith subject,
(i.e., the time of observing \code{yyikj[i,k,j]}).}

\item{nobsik}{number of observations arranged as an integer matrix. \cr
\code{nobsik[i]} is the number of observations for the kth longitudinal measure of the ith subject.}

\item{zzi}{a numeric matrix of covariates in the multistate model. \cr
\code{zzi[i,k]} is the kth covariate of the ith subject.}

\item{xxi}{a numeric matrix of covariates in the longitudinal model. \cr
\code{xxi[i,k]} is the kth covariate of the ith subject.}

\item{statebegin}{an integer matrix of starting states for the event-history intervals. \cr
The states should be encoded as consecutive integers ranging from 1 to the total number of states and 0 represents censoring. \cr
For example, if the event history for the ith subject is \cr
\eqn{[t_1,t_2)} in state \eqn{s_1}, \cr
\eqn{[t_2,t_3)} in state \eqn{s_2}, \cr
\eqn{[t_3,t_4)} in state \eqn{s_3}, \cr
and censored after \eqn{t_4}, then \cr
\code{statebegin[i,1]} is \eqn{s_1}, \cr
\code{statebegin[i,2]} is \eqn{s_2}, \cr
\code{statebegin[i,3]} is \eqn{s_3}, \cr}

\item{stateend}{an integer matrix of destination states for the event-history intervals. \cr
The states should be encoded as consecutive integers ranging from 1 to the total number of states and 0 represents censoring. \cr
For example, if the event history for the ith subject is \cr
\eqn{[t_1,t_2)} in state \eqn{s_1}, \cr
\eqn{[t_2,t_3)} in state \eqn{s_2}, \cr
\eqn{[t_3,t_4)} in state \eqn{s_3}, \cr
and censored after \eqn{t_4}, then \cr
\code{stateend[i,1]} is \eqn{s_2}, \cr
\code{stateend[i,2]} is \eqn{s_3}, \cr
\code{stateend[i,3]} is 0 (0 indicates censoring state). \cr}

\item{timebegin}{an integer matrix of lower bounds for the event-history intervals. \cr
For example, if the event history for the ith subject is \cr
\eqn{[t_1,t_2)} in state \eqn{s_1}, \cr
\eqn{[t_2,t_3)} in state \eqn{s_2}, \cr
\eqn{[t_3,t_4)} in state \eqn{s_3}, \cr
and censored after \eqn{t_4}, then \cr
\code{timebegin[i,1]} is \eqn{t_1}, \cr
\code{timebegin[i,2]} is \eqn{t_2}, \cr
\code{timebegin[i,3]} is \eqn{t_3}. \cr}

\item{timeend}{an integer matrix of upper bounds for the event-history intervals. \cr
For example, if the event history for the ith subject is \cr
\eqn{[t_1,t_2)} in state \eqn{s_1}, \cr
\eqn{[t_2,t_3)} in state \eqn{s_2}, \cr
\eqn{[t_3,t_4)} in state \eqn{s_3}, \cr
and censored after \eqn{t_4}, then \cr
\code{timeend[i,1]} is \eqn{t_2}, \cr
\code{timeend[i,2]} is \eqn{t_3}, \cr
\code{timeend[i,3]} is \eqn{t_4}. \cr}

\item{ntransition}{an integer vector of the number of transitions.
For example, if the event history for the ith subject is \cr
\eqn{[t_1,t_2)} in state \eqn{s_1}, \cr
\eqn{[t_2,t_3)} in state \eqn{s_2}, \cr
\eqn{[t_3,t_4)} in state \eqn{s_3}, \cr
and censored after \eqn{t_4}, then \cr
\code{ntransition[i]} is 3.}

\item{possible_transitions}{a logical matrix indicating the possible transitions. \cr
\code{nrow(possible_transitions)} and \code{ncol(possible_transitions)} should be equal to the total number of states. \cr
\code{possible_transitions[s1,s2]} is \code{TRUE} if the transition from state \code{s1} to \code{s2} is possible. \cr
\code{possible_transitions[s1,s2]} is \code{FALSE} if the transition from state \code{s1} to \code{s2} is not possible. \cr}

\item{basismatrix_long}{a list of basis matrices. \cr
\code{basismatrix_long[[k]]} is the basis matrix for the kth longitudinal measure. \cr
\code{basismatrix_long[[k]][t,l]} is the lth basis function of the kth longitudinal measure evaluated at time \code{t}.}

\item{nqmc}{an integer value. \cr 
Number of quasi Monte Carlo points for evaluating numeric integration.}
}
\value{
a list containing the model fit result.
}
\description{
The function \code{joint_multistate_model} fits 
a joint model of multivariate longitudinal data and multistate data
}
\details{
Fit a joint model of multivariate longitudinal data and multistate data
}
