% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Calculate the log-likelihood}
\usage{
loglikelihood(model_fit, nqmc = 2000)
}
\arguments{
\item{model_fit}{a model fit returned by \code{joint_multistate_model},}

\item{nqmc}{an integer value. \cr 
Number of quasi Monte Carlo points for evaluating numeric integration.}
}
\value{
a list containing the log-likelihood, AIC, and BIC.
}
\description{
Calculate the log-likelihood
}
